fromPrimitive

+++ src/Bookshop/Catalog/Infrastructure/Persistence/Pdo/PdoBookRepository.php
@@ -5,12 +5,8 @@ namespace Bookshop\Catalog\Infrastructure\Persistence\Pdo;
 use PDO;
 use Bookshop\Catalog\Domain\Model\Book\Book;
 use Bookshop\Catalog\Domain\Model\Book\BookId;
-use Bookshop\Catalog\Domain\Model\Book\BookTitle;
 use Bookshop\Catalog\Domain\Model\Book\BookRepository;
 use Bookshop\Catalog\Domain\Model\Genre\Genre;
-use Bookshop\Catalog\Domain\Model\Genre\GenreId;
-use Bookshop\Catalog\Domain\Model\Genre\GenreName;
-use Bookshop\Catalog\Domain\Model\Genre\GenreNumberOfBooks;
 use Exception;
 
 class PdoBookRepository extends PdoRepository implements BookRepository
@@ -22,18 +18,36 @@ class PdoBookRepository extends PdoRepository implements BookRepository
 
     public function all(int $offset, int $limit, string $filter): array
     {
+        $where = empty($filter) ? '' : "WHERE title LIKE '%$filter%'";
+
         $sql = <<<SQL
 SELECT id, title
 FROM books
-WHERE title LIKE "%$filter%"
+$where
 ORDER BY title
 LIMIT $limit OFFSET $offset;
 SQL;
 
         $stmt = $this->connection->prepare($sql);
-        $stmt->execute();
-        $books = $stmt->fetchAll();
-        $bookIds = array_map(fn ($book) => $book['id'], $books);
+        if ($stmt->execute() === false) {
+            throw new Exception('Could not fetch books');
+        }
+
+        $rows = $stmt->fetchAll();
+        if ($rows === false) {
+            throw new Exception('Could not fetch books');
+        } elseif (empty($rows)) {
+            return [];
+        }
+
+        $books = array_map(function ($row) {
+            return Book::fromPrimitives(
+                $row['id'],
+                $row['title'],
+            );
+        }, $rows);
+
+        $bookIds = array_column($rows, 'id');
         $inQuery = str_repeat('?,', count($bookIds) - 1) . '?';
 
         $sql = <<<SQL
@@ -44,24 +58,22 @@ WHERE book_id IN ($inQuery)
 SQL;
 
         $stmt = $this->connection->prepare($sql);
-        $stmt->execute($bookIds);
-        $genres = $stmt->fetchAll();
+        if ($stmt->execute($bookIds) === false) {
+            throw new Exception('Could not fetch genres');
+        }
 
+        $rows = $stmt->fetchAll();
         $genresByBookId = [];
-        foreach ($genres as $genre) {
-            $genresByBookId[$genre['book_id']][] = new Genre(
-                new GenreId($genre['id']),
-                new GenreName($genre['name']),
-                new GenreNumberOfBooks($genre['number_of_books']),
+        foreach ($rows as $row) {
+            $genresByBookId[$row['book_id']][] = Genre::fromPrimitives(
+                $row['id'],
+                $row['name'],
+                $row['number_of_books'],
             );
         }
 
         return array_map(function ($book) use ($genresByBookId) {
-            return new Book(
-                new BookId($book['id']),
-                new BookTitle($book['title']),
-                $genresByBookId[$book['id']] ?? [],
-            );
+            return $book->setGenres($genresByBookId[$book->bookId()->value()]);
         }, $books);
     }
 
@@ -73,12 +85,16 @@ FROM books
 WHERE title LIKE "%$filter%"
 SQL;
         $stmt = $this->connection->prepare($sql);
-        $stmt->execute();
-        $result = $stmt->fetch();
-        if (is_array($result) === false) {
+        if ($stmt->execute() === false) {
             throw new Exception('Could not count books');
         }
-        return (int) $result['total'];
+
+        $row = $stmt->fetch();
+        if ($row === false || is_array($row) === false) {
+            throw new Exception('Could not count books');
+        }
+
+        return (int) $row['total'];
     }
 
     public function ofBookId(BookId $bookId): ?Book
@@ -86,11 +102,19 @@ SQL;
         $sql = "SELECT id, title FROM books WHERE id = :id";
         $stmt = $this->connection->prepare($sql);
         $stmt->bindValue('id', $bookId->value(), PDO::PARAM_STR);
-        $stmt->execute();
-        $book = $stmt->fetch();
-        if (is_array($book) === false) {
+        if ($stmt->execute() === false) {
+            throw new Exception('Could not fetch book');
+        }
+
+        $row = $stmt->fetch();
+        if ($row === false || is_array($row) === false) {
             return null;
         }
+
+        $book = Book::fromPrimitives(
+            $row['id'],
+            $row['title'],
+        );
 
         $sql = <<<SQL
 SELECT id, name, number_of_books
@@ -104,17 +128,23 @@ SQL;
             throw new Exception('Could not count books');
         }
 
-        return new Book(
-            new BookId($book['id']),
-            new BookTitle($book['title']),
-            array_map(function ($genre) {
-                return new Genre(
-                    new GenreId($genre['id']),
-                    new GenreName($genre['name']),
-                    new GenreNumberOfBooks($genre['number_of_books']),
-                );
-            }, $stmt->fetchAll()),
-        );
+        $rows = $stmt->fetchAll();
+
+        if ($rows === false) {
+            throw new Exception('Could not fetch genres');
+        } elseif (empty($rows)) {
+            return $book;
+        }
+
+        $genres = array_map(function ($row) {
+            return Genre::fromPrimitives(
+                $row['id'],
+                $row['name'],
+                $row['number_of_books'],
+            );
+        }, $rows);
+
+        return $book->setGenres($genres);
     }
 
     public function insert(Book $book): void
@@ -126,7 +156,7 @@ SQL;
         $stmt->bindValue('id', $book->bookId()->value(), PDO::PARAM_STR);
         $stmt->bindValue('title', $book->bookTitle()->value(), PDO::PARAM_STR);
 
-        if (!$stmt->execute()) {
+        if ($stmt->execute() === false) {
             $this->connection->rollBack();
             throw new Exception('Could not insert book');
         }
@@ -136,7 +166,7 @@ SQL;
         $stmt->bindValue('book_id', $book->bookId()->value(), PDO::PARAM_STR);
         foreach ($book->bookGenres() as $genre) {
             $stmt->bindValue('genre_id', $genre->genreId()->value(), PDO::PARAM_STR);
-            if (!$stmt->execute()) {
+            if ($stmt->execute() === false) {
                 $this->connection->rollBack();
                 throw new Exception('Could not insert book');
             }
@@ -156,7 +186,7 @@ SQL;
         $stmt->bindValue('id', $book->bookId()->value(), PDO::PARAM_STR);
         $stmt->bindValue('title', $book->bookTitle()->value(), PDO::PARAM_STR);
 
-        if (!$stmt->execute()) {
+        if ($stmt->execute() === false) {
             $this->connection->rollBack();
             throw new Exception('Could not update book');
         }
@@ -165,7 +195,7 @@ SQL;
         $stmt = $this->connection->prepare($sql);
         $stmt->bindValue('book_id', $book->bookId()->value(), PDO::PARAM_STR);
 
-        if (!$stmt->execute()) {
+        if ($stmt->execute() === false) {
             $this->connection->rollBack();
             throw new Exception('Could not update book');
         }
@@ -175,7 +205,7 @@ SQL;
         $stmt->bindValue('book_id', $book->bookId()->value(), PDO::PARAM_STR);
         foreach ($book->bookGenres() as $genre) {
             $stmt->bindValue('genre_id', $genre->genreId()->value(), PDO::PARAM_STR);
-            if (!$stmt->execute()) {
+            if ($stmt->execute() === false) {
                 $this->connection->rollBack();
                 throw new Exception('Could not update book');
             }
@@ -188,15 +218,27 @@ SQL;
 
     public function remove(Book $book): void
     {
+        $this->connection->beginTransaction();
+
         $sql = "DELETE FROM books WHERE id = :id";
         $stmt = $this->connection->prepare($sql);
         $stmt->bindValue('id', $book->bookId()->value(), PDO::PARAM_STR);
-        $stmt->execute();
+
+        if ($stmt->execute() === false) {
+            $this->connection->rollBack();
+            throw new Exception('Could not remove book');
+        }
 
         $sql = "DELETE FROM books_genres WHERE book_id = :book_id";
         $stmt = $this->connection->prepare($sql);
         $stmt->bindValue('book_id', $book->bookId()->value(), PDO::PARAM_STR);
+
         if ($stmt->execute() === false) {
+            $this->connection->rollBack();
+            throw new Exception('Could not remove book');
+        }
+
+        if ($this->connection->commit() === false) {
             throw new Exception('Could not remove book');
         }
     }
